more safer build project tool

+++ tools/build_project.sh
@@ -1,3 +1,7 @@
+if [ $# -eq 0 ]; then
+    echo "No arguments provided"
+    exit 1
+fi
 cd ../$1
 find . -type f ! -name 'CNAME' ! -path './.git/*' -delete
 cp ../tavmant/@prod/* .
